Move cmd signal handler into shared util
