+++ cmd/controller/BUILD.bazel
@@ -25,6 +25,7 @@ go_library(
         "//pkg/issuer/venafi:go_default_library",
         "//pkg/logs:go_default_library",
         "//pkg/util:go_default_library",
+        "//pkg/util/cmd:go_default_library",
         "//pkg/util/feature:go_default_library",
         "@com_github_spf13_cobra//:go_default_library",
         "@io_k8s_apimachinery//pkg/util/errors:go_default_library",
