+++ cmd/webhook/BUILD.bazel
@@ -9,6 +9,7 @@ go_library(
     deps = [
         "//cmd/webhook/app:go_default_library",
         "//cmd/webhook/app/options:go_default_library",
+        "//pkg/util/cmd:go_default_library",
         "@com_github_spf13_pflag//:go_default_library",
         "@io_k8s_klog//:go_default_library",
         "@io_k8s_klog//klogr:go_default_library",
