+++ cmd/webhook/main.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2019 The Jetstack cert-manager contributors.
+Copyright 2020 The Jetstack cert-manager contributors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -19,8 +19,6 @@ package main
 import (
 	goflag "flag"
 	"os"
-	"os/signal"
-	"syscall"
 
 	"github.com/spf13/pflag"
 	"k8s.io/klog"
@@ -28,6 +26,7 @@ import (
 
 	"github.com/jetstack/cert-manager/cmd/webhook/app"
 	"github.com/jetstack/cert-manager/cmd/webhook/app/options"
+	"github.com/jetstack/cert-manager/pkg/util/cmd"
 )
 
 func main() {
@@ -39,32 +38,10 @@ func main() {
 	pflag.Parse()
 
 	log := klogr.New()
-	stopCh := setupSignalHandler()
+	stopCh := cmd.SetupSignalHandler()
 
 	if err := app.RunServer(log, *opts, stopCh); err != nil {
 		log.Error(err, "error running server")
 		os.Exit(1)
 	}
-}
-
-var shutdownSignals = []os.Signal{os.Interrupt, syscall.SIGTERM}
-var onlyOneSignalHandler = make(chan struct{})
-
-// setupSignalHandler registered for SIGTERM and SIGINT. A stop channel is returned
-// which is closed on one of these signals. If a second signal is caught, the program
-// is terminated with exit code 1.
-func setupSignalHandler() (stopCh <-chan struct{}) {
-	close(onlyOneSignalHandler) // panics when called twice
-
-	stop := make(chan struct{})
-	c := make(chan os.Signal, 2)
-	signal.Notify(c, shutdownSignals...)
-	go func() {
-		<-c
-		close(stop)
-		<-c
-		os.Exit(1) // second signal. Exit directly.
-	}()
-
-	return stop
 }
