+++ pkg/util/cmd/signal_windows.go
@@ -14,27 +14,10 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package cmd
 
 import (
-	"flag"
-
-	"k8s.io/klog"
-
-	logf "github.com/jetstack/cert-manager/pkg/logs"
-	"github.com/jetstack/cert-manager/pkg/util/cmd"
+	"os"
 )
 
-func main() {
-	logf.InitLogs(flag.CommandLine)
-	defer logf.FlushLogs()
-
-	stopCh := cmd.SetupSignalHandler()
-	cmd := NewCommandStartCertManagerController(stopCh)
-	cmd.Flags().AddGoFlagSet(flag.CommandLine)
-
-	flag.CommandLine.Parse([]string{})
-	if err := cmd.Execute(); err != nil {
-		klog.Info(err)
-	}
-}
+var shutdownSignals = []os.Signal{os.Interrupt}
